Refactor mdxrp.py to handle jupyter notebooks properly.
Previously, it wouldn't handle refernces to labels that were defined in a subseuent
markdown cell.  I was performig two pass preprocessing: replace labels then replace
refernces, but I was only do it one markdown cell at a time.  So ti would handle
a label defined after its first reference so long as both the label and reference
were in the same markdown cell, but not if the label was in a subsequent markdown cell.
I modifeid the code to do a complete first pass on all markdown cells looking for labels
and then replaces references and labels on all markdown cells in a second pass.

I also made quite a few modifications to satisfy pylint.